test_kafka/requirements.txt
numpy==1.21.1
pandas==1.3.4; python_full_version >= "3.7.1"
pandasql==0.7.3
kafka-python